refactor(services): PR-3.9 reg/builder.py 1108 → analysis.py 抽出
services/reg/builder.py 1108 行里 ~370 行是"看 + 算"性质的纯分析 / 评分 /
搜索过滤函数，跟主流程 (_build_for_subfolder / _build_inner / build) 的
"贪心拉 + 落盘"职责正交。抽到 analysis.py 后 builder.py 瘦到 742 行，主
循环逻辑不再被分析函数夹断。

analysis.py 424 行，含 11 个函数：

  纯分析（builder 主流程用）
    analyze_dataset_structure       扫 train_dir 出 tag 频率 / 分辨率统计
    collect_source_image_ids        扫源数据集 post_id 用于去重
    collect_existing_reg_per_subfolder  增量模式扫已存在 reg
    calculate_tag_similarity        评分：tag 频率向量负 MSE
    calculate_resolution_similarity 评分：aspect 0.6 + resolution 0.4
    calculate_missing_tags          "还差什么 tag" 优先级队列
    check_aspect_ratio              单图长宽比过滤
    find_best_match                 一批 posts 挑最高分

  半公开（builder + tests 用）
    _normalize_tags
    analyze_tags_in_file
    _search_with_filters
    _IMAGE_EXT_NODOT                datasets.IMAGE_EXTS 去点版

兼容：builder.py 顶部 `from .analysis import ...` 把 12 个名 re-export 进本
模块，保 `reg_builder.X` / `from studio.services.reg.builder import X` 旧
import 路径透明工作（tests 大量用 attribute 访问形式，PR-3 shim 经验同此）。
booru_api 的 `monkeypatch.setattr(reg_builder.booru_api, "search_posts", ...)`
patch 改的是共享 module 对象的 attribute，builder 内调用走 module attribute
lookup，不受抽分影响（与 PR-3.8 sources 函数跨边界 patch 失效的场景不同）。

builder.py 内 `from .. import tagedit` import 已无 callsite，顺手删；
analysis.py 自带 `from .. import tagedit` 给 analyze_tags_in_file 用。

回归：tests/test_reg_builder.py + PR-1 安全网 163/163 全绿；全套 25 fail
全 pre-existing（stash 比对 baseline 相同 fail 集，与本 PR 无关）。
